Init-ing git in pre-existing project.
I have worked on this project for weeks. This is the client-side of a mobile app. The desktop version was very tied to the platform due to tight coupling between the platform behaviour and networking components of my program.

This edition is a proof of the challenges I went through, primarily: Making a mobile app, enabling networking on it, making the client-side networking as abstract as possible, decreasing coupling between the networking-component and the behaviour.